Sign-extend retained Milan reference words before averaging

+++ drivers/goodix53x5/milan/preprocess/classification.c
@@ -1237,7 +1237,7 @@ milan_profile9_update_history (GoodixMilanPreprocessState *state,
 
         state->profile9_history_reference[i] =
           (uint16_t) ((normalized +
-                       (int) state->profile9_history_reference[i] *
+                       (int16_t) state->profile9_history_reference[i] *
                          (int) state->profile9_history_count) /
                       (int) (state->profile9_history_count + 1));
       }
